gempa/requirements.txt
python-socketio[client]
websocket-client
pandas
openpyxl
requests